+++ tests/Unit/AssetsTest.php
@@ -92,7 +92,7 @@ final class AssetsTest extends TestCase {
                 $assets->enqueue_admin();
                 $this->assertContains( 'fbm-admin', $fbm_enqueued_styles );
 
-                $current_screen       = (object) array( 'id' => 'foodbank_page_fbm_settings' );
+                $current_screen       = (object) array( 'id' => 'foodbank_page_fbm_database' );
                 $fbm_enqueued_styles = array();
                 $assets->enqueue_admin();
                 $this->assertContains( 'fbm-admin', $fbm_enqueued_styles );